2023 - Nim: Set up framework
Gonna be pretty busy this December. Choosing Nim as it's
procedural and familiar. Not sure if I'll see any brand new language
features this year.

The first 3 days from 2019 have been implemented for a primer on Nim
and to get the framework into a good shape. To delete as the 2023 puzzles come out